requirements.txt
pandas==2.2.0
yfinance==0.2.36
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pyyaml==6.0.1
python-dotenv==1.0.1
pytest==8.0.0
